🚀 feat(App.tsx): update import paths for Linka and Router components 🚀 feat(api): update import path for getConfig utility function 🚀 feat(api): move linkding API functions to their own directory 🚀 feat(api/linkding): add doArchive, doAuth, doCreate, doDelete, doUpdate, getBookmark, getBookmarks, and getTags functions 🚀 feat(api/openai): update import path for getConfig utility function 🚀 feat(AddBookmark.tsx): add AddBookmark component The changes update import paths for the Linka and Router components, and the getConfig utility function. The linkding API functions are moved to their own directory and new functions are added. The AddBookmark component is also added.
🎉 feat: add new components to the project
This commit adds new components to the project, including AddBookmark, BookmarkForm, BookmarkFormFillButton, Credits, and Linka. AddBookmark and BookmarkFormFillButton are used to add bookmarks to the application, while BookmarkForm is a form used to fill in bookmark details. Credits is a component that displays the version of the application, and Linka is the main component that wraps the entire application.

🚀 feat(LinkaItem): add LinkaItem component
🚀 feat(LinkaItem): add LinkaItem component to render a bookmark item
🚀 feat(LinkaItem): add LinkaItem component to render a bookmark item with delete and update functionality
🚀 feat(LinkaItem): add LinkaItem component to render a bookmark item with delete and update functionality and drawer
🚀 feat(LinkaItem): add LinkaItem component to render a bookmark item with delete and update functionality and drawer with form
🚀 feat(LinkaItem): add LinkaItem component to render a bookmark item with delete and update functionality and drawer with form and loading icon
🚀 feat(LinkaItem): add LinkaItem component to render a bookmark item with delete and update functionality and drawer with form and loading icon and shortenURL
🚀 feat(LinkaItem): add LinkaItem component to render a bookmark item with delete

🎨 style(components): use named exports instead of default exports
This commit changes the exports of the components to named exports instead of default exports. This improves the readability of the code and makes it easier to import specific components.

🚀 feat(components): add Settings component
This commit adds the Settings component to the exports of the components. The Settings component is used to configure the application settings.

🚚 chore(Auth.tsx): move Auth.tsx to Auth/Auth.tsx
This commit moves the Auth.tsx file to Auth/Auth.tsx. This improves the organization of the code.

🚚 chore(AuthProps.ts): move AuthProps.ts to Auth/AuthProps.ts
This commit moves the AuthProps.ts file to Auth/AuthProps.ts. This improves the organization of the code.

🚀 feat(Auth.tsx): add FormContainer and FormPartLinkdingSettings
This commit adds the FormContainer and FormPartLinkdingSettings components to the Auth.tsx

🎉 feat(utils): add getConfig, setConfig, and shortenURL functions
The getConfig function retrieves the LinkaSettings object from localStorage and merges it with the legacy authObject. The setConfig function updates the LinkaSettings object in localStorage. The shortenURL function removes the 'https://' and 'http://' prefixes from a URL string. These functions improve the maintainability and readability of the codebase.
