+++ src/api/index.ts
@@ -1,8 +1,8 @@
-export { default as doArchive } from './doArchive';
-export { default as doAuth } from './doAuth';
-export { default as doCreate } from './doCreate';
-export { default as doDelete } from './doDelete';
-export { default as doUpdate } from './doUpdate';
-export { default as getBookmark } from './getBookmark';
-export { default as getBookmarks } from './getBookmarks';
-export { default as getTags } from './getTags';
+export { default as doArchive } from './linkding/doArchive';
+export { default as doAuth } from './linkding/doAuth';
+export { default as doCreate } from './linkding/doCreate';
+export { default as doDelete } from './linkding/doDelete';
+export { default as doUpdate } from './linkding/doUpdate';
+export { default as getBookmark } from './linkding/getBookmark';
+export { default as getBookmarks } from './linkding/getBookmarks';
+export { default as getTags } from './linkding/getTags';
