+++ src/api/linkding/doArchive.ts
@@ -1,22 +1,18 @@
-import { bookmarkArgs } from '@/types';
-import { getConfig } from '@/utils/getConfig';
+import { doArgs } from '@/types';
+import { getConfig } from '@/utils/getConfig/getConfig';
 
-export async function doUpdate(args: bookmarkArgs) {
+export async function doArchive(args: doArgs) {
   const config = getConfig();
-
   if (config.token && config.url) {
-    const endpoint = `${config.url}/api/bookmarks/${args.id}/`;
+    const endpoint = `${config.url}/api/bookmarks/${args.id}/archive/`;
     let url = new URL(endpoint);
 
     const res = await fetch(url, {
-      method: 'PUT',
+      method: 'POST',
       headers: {
         Authorization: `Token ${config.token}`,
-        'Content-Type': 'application/json',
       },
-      body: JSON.stringify(args.payload),
     });
-    console.log(res);
     if (res.ok) {
       return true;
     }
@@ -26,4 +22,4 @@ export async function doUpdate(args: bookmarkArgs) {
   }
 }
 
-export default doUpdate;
+export default doArchive;
