+++ src/api/linkding/doDelete.ts
@@ -1,21 +1,18 @@
 import { doArgs } from '@/types';
-import { getConfig } from '@/utils/getConfig';
+import { getConfig } from '@/utils/getConfig/getConfig';
 
-export async function doShare(args: doArgs) {
+export async function doDelete(args: doArgs) {
   const config = getConfig();
   if (config.token && config.url) {
     const endpoint = `${config.url}/api/bookmarks/${args.id}/`;
     let url = new URL(endpoint);
 
     const res = await fetch(url, {
-      method: 'PATCH',
+      method: 'DELETE',
       headers: {
         Authorization: `Token ${config.token}`,
-        'Content-Type': 'application/json',
       },
-      body: JSON.stringify({ shared: true }),
     });
-    console.log(res);
     if (res.ok) {
       return true;
     }
@@ -25,4 +22,4 @@ export async function doShare(args: doArgs) {
   }
 }
 
-export default doShare;
+export default doDelete;
