+++ src/components/AddBookmark/index.ts
@@ -1,0 +1,1 @@
+export { default as AddBookmark } from './AddBookmark';
