+++ src/components/BookmarkForm/index.ts
@@ -1,0 +1,1 @@
+export { default as BookmarkForm } from './BookmarkForm';
