+++ src/components/Credits/index.ts
@@ -1,0 +1,1 @@
+export { default as Credits } from './Credits';
