+++ src/components/Linka/index.ts
@@ -1,0 +1,1 @@
+export { default as Linka } from './Linka';
