+++ src/components/LinkaItem/index.ts
@@ -1,0 +1,1 @@
+export { default as LinkaItem } from './LinkaItem';
