+++ src/components/LoadingIcon/index.ts
@@ -1,0 +1,1 @@
+export { default as LoadingIcon } from './LoadingIcon';
