+++ src/components/Settings/index.ts
@@ -1,1 +1,2 @@
+export { default as FormPartLinkdingSettings } from './FormPartLinkdingSettings';
 export { default as Settings } from './Settings';
