+++ src/pages/Auth/AuthProps.ts
@@ -1,0 +1,5 @@
+export interface AuthProps {
+  setReady: (e: any) => void;
+}
+
+export default AuthProps;
