+++ src/pages/Auth/index.ts
@@ -1,0 +1,2 @@
+export { default as Auth } from './Auth';
+export * from './AuthProps';
