+++ src/utils/getConfig/index.ts
@@ -1,0 +1,1 @@
+export { getConfig } from './getConfig';
