+++ src/utils/index.ts
@@ -1,2 +1,3 @@
-export { getConfig } from './getConfig';
-export { setConfig } from './setConfig';
+export { getConfig } from './getConfig/getConfig';
+export { setConfig } from './setConfig/setConfig';
+export { shortenURL } from './shortenURL/shortenURL';
