+++ src/utils/shortenURL/index.ts
@@ -1,0 +1,1 @@
+export { shortenURL } from './shortenURL';
